fix(scripts): stop the reference parser from dropping documentation
Two defects, both silent:

A JSDoc block without an @element line dropped every @attr, @slot and @fires
it had. Tags attach to the current element and the fallback was only created
after the loop, so nldd-top-title-bar was published with no attributes while
its JSDoc documented six.

"[name]- description" without a space before the separator produced the
literal attribute name "name]-" in the published reference.

The parsing moves to scripts/lib/component-jsdoc.js so the new API check can
share it. Two parsers would let that check pass while the reference says
something else. Verified by regenerating reference.md byte for byte.
